Rename 'natevents' package to 'event'
