+++ nat/event/events_tracker.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package natevents
+package event
 
 import (
 	"time"
