+++ nat/mapping/port_mapping.go
@@ -22,7 +22,7 @@ import (
 
 	log "github.com/cihub/seelog"
 	portmap "github.com/ethereum/go-ethereum/p2p/nat"
-	"github.com/mysteriumnetwork/node/nat/natevents"
+	"github.com/mysteriumnetwork/node/nat/event"
 )
 
 const logPrefix = "[port mapping] "
@@ -88,11 +88,11 @@ func addMapping(m portmap.Interface, protocol string, extPort, intPort int, name
 		log.Debugf("%s Couldn't add port mapping for port %d: %v, retrying with permanent lease", logPrefix, extPort, err)
 		if err := m.AddMapping(protocol, extPort, intPort, name, 0); err != nil {
 			// some gateways support only permanent leases
-			publisher.Publish(natevents.EventTopic, natevents.BuildFailureEvent(StageName, err))
+			publisher.Publish(event.EventTopic, event.BuildFailureEvent(StageName, err))
 			log.Debugf("%s Couldn't add port mapping for port %d: %v", logPrefix, extPort, err)
 			return
 		}
 	}
-	publisher.Publish(natevents.EventTopic, natevents.BuildSuccessEvent(StageName))
+	publisher.Publish(event.EventTopic, event.BuildSuccessEvent(StageName))
 	log.Info(logPrefix, "Mapped network port: ", extPort)
 }
