+++ nat/status_tracker_test.go
@@ -20,7 +20,7 @@ package nat
 import (
 	"testing"
 
-	"github.com/mysteriumnetwork/node/nat/natevents"
+	"github.com/mysteriumnetwork/node/nat/event"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 )
@@ -35,7 +35,7 @@ func Test_StatusTracker_Status_ReturnsNotFinishedInitially(t *testing.T) {
 
 func Test_StatusTracker_Status_ReturnsSuccessful_WithSuccessfulEvent(t *testing.T) {
 	tracker := NewStatusTracker("last stage")
-	tracker.ConsumeNATEvent(natevents.Event{Successful: true, Stage: "any stage"})
+	tracker.ConsumeNATEvent(event.Event{Successful: true, Stage: "any stage"})
 	status := tracker.Status()
 
 	assert.Equal(t, "successful", status.Status)
@@ -44,7 +44,7 @@ func Test_StatusTracker_Status_ReturnsSuccessful_WithSuccessfulEvent(t *testing.
 
 func Test_StatusTracker_Status_ReturnsFailure_WithHolepunchingFailureEvent(t *testing.T) {
 	tracker := NewStatusTracker("last stage")
-	tracker.ConsumeNATEvent(natevents.Event{Successful: false, Stage: "last stage", Error: errors.New("test error")})
+	tracker.ConsumeNATEvent(event.Event{Successful: false, Stage: "last stage", Error: errors.New("test error")})
 	status := tracker.Status()
 
 	assert.Equal(t, "failure", status.Status)
@@ -53,7 +53,7 @@ func Test_StatusTracker_Status_ReturnsFailure_WithHolepunchingFailureEvent(t *te
 
 func Test_StatusTracker_Status_ReturnsNotFinished_WithPortMappingFailureEvent(t *testing.T) {
 	tracker := NewStatusTracker("last stage")
-	tracker.ConsumeNATEvent(natevents.Event{Successful: false, Stage: "first stage"})
+	tracker.ConsumeNATEvent(event.Event{Successful: false, Stage: "first stage"})
 	status := tracker.Status()
 
 	assert.Equal(t, "not_finished", status.Status)
