+++ nat/traversal/pinger.go
@@ -26,7 +26,7 @@ import (
 
 	log "github.com/cihub/seelog"
 	"github.com/mysteriumnetwork/node/core/port"
-	"github.com/mysteriumnetwork/node/nat/natevents"
+	"github.com/mysteriumnetwork/node/nat/event"
 	"github.com/mysteriumnetwork/node/services"
 	"github.com/pkg/errors"
 	"golang.org/x/net/ipv4"
@@ -55,7 +55,7 @@ type Pinger struct {
 
 // NatEventWaiter is responsible for waiting for nat events
 type NatEventWaiter interface {
-	WaitForEvent() natevents.Event
+	WaitForEvent() event.Event
 }
 
 // ConfigParser is able to parse a config from given raw json
@@ -227,11 +227,11 @@ func (p *Pinger) ping(conn *net.UDPConn) error {
 
 			err := p.sendPingRequest(conn, n)
 			if err != nil {
-				p.eventPublisher.Publish(natevents.EventTopic, natevents.BuildFailureEvent(StageName, err))
+				p.eventPublisher.Publish(event.EventTopic, event.BuildFailureEvent(StageName, err))
 				return err
 			}
 
-			p.eventPublisher.Publish(natevents.EventTopic, natevents.BuildSuccessEvent(StageName))
+			p.eventPublisher.Publish(event.EventTopic, event.BuildSuccessEvent(StageName))
 
 			n++
 		}
